backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==10.3.0
sqlparse==0.5.0
